+++ macos_pkg_builder/__main__.py
@@ -1,0 +1,8 @@
+"""
+__main__.py: Validates project
+"""
+
+from . import validation
+
+if __name__ == "__main__":
+    validation.main()